Sigh. It's 2017 and yet here I am on a rainy Saturday afternoon fixing LOG_TRACE code because we are not compiling in this mode in our pipeline

+++ test/type/boolean_value_test.cpp
@@ -111,7 +111,9 @@ TEST_F(BooleanValueTests, ComparisonTest) {
         }  // SWITCH
         LOG_TRACE("%s %s %s => %d | %d\n", val0.ToString().c_str(),
                   ExpressionTypeToString(etype).c_str(),
-                  val1.ToString().c_str(), expected, result);
+                  val1.ToString().c_str(),
+                  static_cast<int>(expected),
+                  static_cast<int>(result));
 
         if (expected_null) expected = false;
